modified:   Ceng477/HW2/.vscode/settings.json modified:   Ceng477/HW2/code_template/code_template/Helpers.cpp modified:   Ceng477/HW2/code_template/code_template/Scene.cpp modified:   Ceng477/HW2/code_template/code_template/empty_box_1.ppm.png new file:   Ceng477/HW2/code_template/code_template/empty_box_2.ppm.png new file:   Ceng477/HW2/code_template/code_template/empty_box_3.ppm.png new file:   Ceng477/HW2/code_template/code_template/empty_box_4.ppm.png new file:   Ceng477/HW2/code_template/code_template/empty_box_5.ppm.png new file:   Ceng477/HW2/code_template/code_template/empty_box_6.ppm.png new file:   Ceng477/HW2/code_template/code_template/empty_box_7.ppm.png new file:   Ceng477/HW2/code_template/code_template/empty_box_8.ppm.png new file:   Ceng477/HW2/code_template/code_template/empty_box_example.ppm.png modified:   Ceng477/HW2/code_template/code_template/rasterizer deleted:    Ceng477/HW2/code_template/code_template/triangle.ppm.png

+++ Ceng477/HW2/code_template/code_template/Helpers.cpp
@@ -215,7 +215,7 @@ Color multiplyColor(Color c, double s){
 * Round the given Color c and return the rounded color.
 */
 Color roundColor(Color c){
-    double r, g, b;
+    int r, g, b;
     r = (int)(c.r + 0.5);
     g = (int)(c.g + 0.5);
     b = (int)(c.b + 0.5);
